various bug fixes, further debugging
Still very much a wip but now it's printing out the
packed bytes it's storing.  For the most part it looks good.
There's a nocall entry being missed in the test data that I'm
using but the bulk of the information looks good.

Offset and tile position arrays need to be verified.